This a partial commit.
The opencv file does not work at the moment because the school's
computers block opencv from running properly.

We used cmder to get around the block of the school's
restrictions. We need to add the second half of the java folder
to get the code to work with eclipse.

We went on to opencv site and and downloaded the proper version to work.
We used the opencv version 2.4.11 for windows because it has certain commands that can work for
taking pictures. Also this older version is more stable than the newer version.