Added local dependency to OpenCV3.1
* lib: New folder, for local libraries.  Having this in Git prevents us
from having to import from an external folder all the time.
* lib\OpenCV31: Downloaded from the OpenCV source code distribution, version 3.1 (http://opencv.org/downloads.html) We
only kept the build/ folders (>300 MB) and deleted the source/ folders
(<150 MB.)  Yes, the next person who does a git pull is about to get a
rude awakening.  Them's the breaks; we'll clean up later.
* .classpath: Modified to reflect Eclipse's new dependency location.
For some reason, it's using an absolute path reference for the
OpenCV-310.jar file.  That makes us sad.  We wanted that reference to be
a relative path.  Will fix in a subsequent commit if possible.

We want to avoid a situation where everyone who is using this code ends
up having to add an external JAR file to their Eclipse instead of having
the project manage that automatically.