Added a Camera Thread and spawned it in robotInit
CameraRunnable.java: We implemented a thread runnable. All it does right now is sleep and
print a message. We implemented some thread killing code and we hope it works...

Robot.java: We spawned the cameraRunnable. We added a thread finalizer that should set the death flag to kill the thread.