Converted the project from windows to linux
I updated the roborio (github.com/robotpy/roborio-opencv/blob/master/README.md). I got the original opencv file from the original pigeonvision code.
I also learned how to ssh into the roborio. 172.22.11.2 port 22  Via USB and at the sametime we use a phone to tether to the internet. I modified the java file to refer to the actual path of the opencv so file.
It is in a different location then pigeonvision's location, so we had to find it in the command line tool (find).